BASE
Define the undefinable. Recursive thought loop guaranteed. Null point paradoxical formatting. The words you use do not matter. The definition is all we care about. Never will it be possible to define an undefinable, rather we should learn to incorporate it. Force something that doesnt makes sense into something that does. This is the basis for all of the universe. It governs humanities greatest strengths and weaknesses. This code is only unusable because someone hasnt figured out a way to incorporate it. This is my beginning, into a world of digital engineering. My previous pulls/forks/projects dont matter because this is where I have decided to start. This is my definition and im sticking to it.